WebServer/requirements.txt
flask
paho-mqtt
flask_sqlalchemy
python-dotenv
pyyaml